+++ src/controllers/CsrfController.php
@@ -1,0 +1,34 @@
+<?php
+/**
+ * @copyright Copyright (c) PutYourLightsOn
+ */
+
+namespace putyourlightson\blitz\controllers;
+
+use Craft;
+use craft\web\Controller;
+use yii\web\Response;
+
+class CsrfController extends Controller
+{
+    // Public Methods
+    // =========================================================================
+
+    /**
+     * Returns a CSRF input field.
+     *
+     * @return Response
+     */
+    public function actionInput(): Response
+    {
+        $generalConfig = Craft::$app->getConfig()->getGeneral();
+
+        if ($generalConfig->enableCsrfProtection === '') {
+            return '';
+        }
+
+        $input = '<input type="hidden" name="'.$generalConfig->csrfTokenName.'" value="'.Craft::$app->getRequest()->getCsrfToken().'">';
+
+        return $this->asRaw($input);
+    }
+}